+++ backend/core/src/auth/dto/user-profile.dto.ts
@@ -3,16 +3,19 @@ import { User } from "../entities/user.entity"
 import { Expose, Type } from "class-transformer"
 import {
   IsDefined,
+  IsEmail,
   IsNotEmpty,
   IsOptional,
   IsString,
   Matches,
+  MaxLength,
   ValidateIf,
   ValidateNested,
 } from "class-validator"
 import { ValidationsGroupsEnum } from "../../shared/types/validations-groups-enum"
 import { passwordRegex } from "../../shared/password-regex"
 import { IdDto } from "../../shared/dto/id.dto"
+import { EnforceLowerCase } from "../../shared/decorators/enforceLowerCase.decorator"
 
 export class UserProfileUpdateDto extends PickType(User, [
   "id",
@@ -23,6 +26,7 @@ export class UserProfileUpdateDto extends PickType(User, [
   "createdAt",
   "updatedAt",
   "language",
+  "phoneNumber",
 ] as const) {
   @Expose()
   @IsOptional({ groups: [ValidationsGroupsEnum.default] })
@@ -43,4 +47,16 @@ export class UserProfileUpdateDto extends PickType(User, [
   @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
   @Type(() => IdDto)
   jurisdictions: IdDto[]
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsEmail({}, { groups: [ValidationsGroupsEnum.default] })
+  @EnforceLowerCase()
+  newEmail?: string
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @MaxLength(256, { groups: [ValidationsGroupsEnum.default] })
+  appUrl?: string | null
 }