+++ backend/core/src/auth/services/user.service.ts
@@ -140,6 +140,15 @@ export class UserService {
       delete dto.jurisdictions
     }
 
+    if (dto.newEmail && dto.appUrl) {
+      user.confirmationToken = UserService.createConfirmationToken(user.id, dto.newEmail)
+      const confirmationUrl = UserService.getPublicConfirmationUrl(dto.appUrl, user)
+      await this.emailService.changeEmail(user, dto.appUrl, confirmationUrl, dto.newEmail)
+    }
+
+    delete dto.newEmail
+    delete dto.appUrl
+
     assignDefined(user, {
       ...dto,
       passwordHash,
@@ -157,14 +166,6 @@ export class UserService {
       throw new HttpException(USER_ERRORS.NOT_FOUND.message, USER_ERRORS.NOT_FOUND.status)
     }
 
-    if (user.confirmedAt) {
-      console.error(`User ${token.id} already confirmed.`)
-      throw new HttpException(
-        USER_ERRORS.ACCOUNT_CONFIRMED.message,
-        USER_ERRORS.ACCOUNT_CONFIRMED.status
-      )
-    }
-
     if (user.confirmationToken !== dto.token) {
       console.error(`Confirmation token mismatch for user ${token.id}.`)
       throw new HttpException(USER_ERRORS.TOKEN_MISSING.message, USER_ERRORS.TOKEN_MISSING.status)
@@ -178,14 +179,26 @@ export class UserService {
     }
 
     try {
-      await this.userRepository.save(user)
+      await this.userRepository.save({
+        ...user,
+        ...(token.email && { email: token.email }),
+      })
       return this.authService.generateAccessToken(user)
     } catch (err) {
       throw new HttpException(USER_ERRORS.ERROR_SAVING.message, USER_ERRORS.ERROR_SAVING.status)
     }
   }
 
-  public async resendConfirmation(dto: EmailDto) {
+  private static createConfirmationToken(userId: string, email: string) {
+    const payload = {
+      id: userId,
+      email,
+      exp: Number.parseInt(moment().add(24, "hours").format("X")),
+    }
+    return encode(payload, process.env.APP_SECRET)
+  }
+
+  public async resendPublicConfirmation(dto: EmailDto) {
     const user = await this.findByEmail(dto.email)
     if (!user) {
       throw new HttpException(USER_ERRORS.NOT_FOUND.message, USER_ERRORS.NOT_FOUND.status)
@@ -196,8 +209,7 @@ export class UserService {
         USER_ERRORS.ACCOUNT_CONFIRMED.status
       )
     } else {
-      const payload = { id: user.id, exp: Number.parseInt(moment().add(24, "hours").format("X")) }
-      user.confirmationToken = encode(payload, process.env.APP_SECRET)
+      user.confirmationToken = UserService.createConfirmationToken(user.id, user.email)
       try {
         await this.userRepository.save(user)
         const confirmationUrl = UserService.getPublicConfirmationUrl(dto.appUrl, user)
@@ -246,11 +258,7 @@ export class UserService {
     try {
       let newUser = await this.userRepository.save(dto)
 
-      const payload = {
-        id: newUser.id,
-        expiresAt: Number.parseInt(moment().add(24, "hours").format("X")),
-      }
-      newUser.confirmationToken = encode(payload, process.env.APP_SECRET)
+      newUser.confirmationToken = UserService.createConfirmationToken(newUser.id, newUser.email)
       newUser = await this.userRepository.save(newUser)
 
       return newUser
@@ -260,7 +268,11 @@ export class UserService {
     }
   }
 
-  public async createUser(dto: UserCreateDto, authContext: AuthContext, sendWelcomeEmail = false) {
+  public async createPublicUser(
+    dto: UserCreateDto,
+    authContext: AuthContext,
+    sendWelcomeEmail = false
+  ) {
     const newUser = await this._createUser(
       {
         ...dto,
@@ -318,7 +330,7 @@ export class UserService {
     return qb
   }
 
-  async invite(dto: UserInviteDto, authContext: AuthContext) {
+  async invitePartnersPortalUser(dto: UserInviteDto, authContext: AuthContext) {
     const password = crypto.randomBytes(8).toString("hex")
     const user = await this._createUser(
       {