+++ backend/core/types/src/backend-swagger.ts
@@ -1656,6 +1656,116 @@ export class PropertyGroupsService {
   }
 }
 
+export class ProgramsService {
+  /**
+   * List preferences
+   */
+  list(
+    params: {
+      /**  */
+      filter?: ProgramsFilterParams[]
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<Program[]> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/programs"
+
+      const configs: IRequestConfig = getConfigs("get", "application/json", url, options)
+      configs.params = { filter: params["filter"] }
+      let data = null
+
+      configs.data = data
+      axios(configs, resolve, reject)
+    })
+  }
+  /**
+   * Create program
+   */
+  create(
+    params: {
+      /** requestBody */
+      body?: ProgramCreate
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<Program> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/programs"
+
+      const configs: IRequestConfig = getConfigs("post", "application/json", url, options)
+
+      let data = params.body
+
+      configs.data = data
+      axios(configs, resolve, reject)
+    })
+  }
+  /**
+   * Update program
+   */
+  update(
+    params: {
+      /** requestBody */
+      body?: ProgramUpdate
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<Program> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/programs/{programId}"
+
+      const configs: IRequestConfig = getConfigs("put", "application/json", url, options)
+
+      let data = params.body
+
+      configs.data = data
+      axios(configs, resolve, reject)
+    })
+  }
+  /**
+   * Get program by id
+   */
+  retrieve(
+    params: {
+      /**  */
+      programId: string
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<Program> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/programs/{programId}"
+      url = url.replace("{programId}", params["programId"] + "")
+
+      const configs: IRequestConfig = getConfigs("get", "application/json", url, options)
+
+      let data = null
+
+      configs.data = data
+      axios(configs, resolve, reject)
+    })
+  }
+  /**
+   * Delete program by id
+   */
+  delete(
+    params: {
+      /**  */
+      programId: string
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<any> {
+    return new Promise((resolve, reject) => {
+      let url = basePath + "/programs/{programId}"
+      url = url.replace("{programId}", params["programId"] + "")
+
+      const configs: IRequestConfig = getConfigs("delete", "application/json", url, options)
+
+      let data = null
+
+      configs.data = data
+      axios(configs, resolve, reject)
+    })
+  }
+}
+
 export class ReservedCommunityTypesService {
   /**
    * List reservedCommunityTypes
@@ -3784,6 +3894,9 @@ export interface User {
   dob?: Date
 
   /**  */
+  phoneNumber?: string
+
+  /**  */
   createdAt: Date
 
   /**  */
@@ -3826,6 +3939,9 @@ export interface UserCreate {
 
   /**  */
   dob?: Date
+
+  /**  */
+  phoneNumber?: string
 }
 
 export interface UserBasic {
@@ -3861,6 +3977,9 @@ export interface UserBasic {
 
   /**  */
   dob?: Date
+
+  /**  */
+  phoneNumber?: string
 
   /**  */
   createdAt: Date
@@ -3943,6 +4062,12 @@ export interface UserUpdate {
   leasingAgentInListings?: Id[]
 
   /**  */
+  newEmail?: string
+
+  /**  */
+  appUrl?: string
+
+  /**  */
   confirmedAt?: Date
 
   /**  */
@@ -3956,6 +4081,9 @@ export interface UserUpdate {
 
   /**  */
   dob?: Date
+
+  /**  */
+  phoneNumber?: string
 }
 
 export interface UserFilterParams {
@@ -4015,6 +4143,9 @@ export interface UserInvite {
 
   /**  */
   dob?: Date
+
+  /**  */
+  phoneNumber?: string
 }
 
 export interface UserProfileUpdate {
@@ -4029,6 +4160,12 @@ export interface UserProfileUpdate {
 
   /**  */
   jurisdictions: Id[]
+
+  /**  */
+  newEmail?: string
+
+  /**  */
+  appUrl?: string
 
   /**  */
   id: string
@@ -5893,6 +6030,45 @@ export interface PropertyGroupUpdate {
   id: string
 }
 
+export interface ProgramsFilterParams {
+  /**  */
+  $comparison: EnumProgramsFilterParamsComparison
+
+  /**  */
+  jurisdiction?: string
+}
+
+export interface ProgramCreate {
+  /**  */
+  title?: string
+
+  /**  */
+  subtitle?: string
+
+  /**  */
+  description?: string
+
+  /**  */
+  formMetadata?: FormMetadata
+}
+
+export interface ProgramUpdate {
+  /**  */
+  title?: string
+
+  /**  */
+  subtitle?: string
+
+  /**  */
+  description?: string
+
+  /**  */
+  formMetadata?: FormMetadata
+
+  /**  */
+  id: string
+}
+
 export interface ReservedCommunityTypeCreate {
   /**  */
   jurisdiction: Id
@@ -6148,6 +6324,13 @@ export type CombinedLeasingAgentAddressTypes = AddressUpdate
 export type CombinedResultTypes = AssetCreate
 export type CombinedBuildingAddressTypes = AddressUpdate
 export enum EnumPreferencesFilterParamsComparison {
+  "=" = "=",
+  "<>" = "<>",
+  "IN" = "IN",
+  ">=" = ">=",
+  "NA" = "NA",
+}
+export enum EnumProgramsFilterParamsComparison {
   "=" = "=",
   "<>" = "<>",
   "IN" = "IN",