+++ ui-components/src/forms/DOBField.tsx
@@ -33,8 +33,8 @@ const DOBField = (props: DOBFieldProps) => {
     return [name, baseName].filter((item) => item).join(".")
   }
 
-  const birthDay = watch(getFieldName("birthDay"))
-  const birthMonth = watch(getFieldName("birthMonth"))
+  const birthDay = watch(getFieldName("birthDay")) ?? defaultDOB?.birthDay
+  const birthMonth = watch(getFieldName("birthMonth")) ?? defaultDOB?.birthMonth
 
   const validateAge = (value: string) => {
     return (